fix: harden prototype runtime — auth, security, SDK alignment, remove sim
- Auth P0: SessionManager::create_session stored empty roles; RBAC always
  403'd admin users. Add create_session_with_roles() and bootstrap admin/
  editor/viewer roles + assign admin at daemon startup (novad, session,
  manager, middleware).
- Security: enforce body size limit (10MB RequestBodyLimitLayer), API key
  RBAC on mutating /auth endpoints, downgrade 200 request logs to debug,
  require truncate on SSTable creation, add safety docs for unsafe fns.
- SDK: realign auth.ts/types.ts/client.ts with actual REST surface
  (/auth/login, /auth/refresh, /auth/users, /auth/api-keys), fix
  RefreshAuthProvider endpoint, remove stale Role/Connection exports.
- Remove dead nova-sim crate (TUI sim not needed for prototypes), drop
  workspace member and CI/Makefile excludes.
- Clippy -D warnings clean: fix collapsible_if, len_zero, manual_strip,
  explicit_counter_loop, needless_range_loop, suspicious_open_options,
  type_complexity, result_large_err, while_let_loop across 15+ crates;
  cargo fmt --all. CI and SDK/dashboard tsc now green.

+++ crates/nova-api/src/admin.rs
@@ -155,9 +155,7 @@ async fn health_check(State(state): State<Arc<AdminState>>) -> Json<Value> {
                 .load(std::sync::atomic::Ordering::Relaxed) as u64
         })
         .unwrap_or(0);
-    let connections_active = (pipeline_snap.active_operations as u64)
-        .max(ws_subscribers)
-        .max(1); // at least this health request
+    let connections_active = pipeline_snap.active_operations.max(ws_subscribers).max(1); // at least this health request
     let request_rate = pipeline_snap.operations_total.saturating_sub(0) / uptime.max(1);
 
     let subsystems = json!({
